Fix issue with locating textarea among multiple textareas
Previously if there were multiple textareas the selection logic
failed if the textarea were empty or already fully selected. As
a workaround, any such areas are temporarily set to an a
single space (' '). This gives something for selectAll to work on
(or clears an active selection). After locating the textarea, we
then restore the previously value.

Signed-off-by: Jay Soffian <jaysoffian@gmail.com>


git-svn-id: http://svn.textmate.org/trunk/Tools/Edit%20in%20TextMate@8982 dfb7d73b-c2ec-0310-8fea-fb051d288c6d